\name{ddesolve-package}
\alias{ddesolve-package}
\alias{ddesolve}
\docType{package}
\title{
Solver for Delay Differential Equations
}
\description{
A solver for systems of delay differential equations based off numerical routines from Simon Wood's solv95 program. 
This solver is also capable of solving systems of ordinary differential equations.
}
\details{
\tabular{ll}{
Package: \tab ddesolve\cr
Type: \tab Package\cr
Version: \tab 1.02\cr
Date: \tab 2008-04-17\cr
License: \tab GNU GENERAL PUBLIC LICENSE Version 2\cr
}
Please see the user guide vignette, located in ddesolve/doc, for a comprehensive overview.
}

\author{
Alex Couture-Beil <ddesolve@mofo.ca>, 
  Jon T. Schnute <SchnuteJ@pac.dfo-mpo.gc.ca>,
  Rowan Haigh <HaighR@pac.dfo-mpo.gc.ca>

Maintainer: Alex Couture-Beil <ddesolve@mofo.ca>
}

\references{
Wood, S.N. 1999. Solv95: a numerical solver for systems of delay differential equations with
switches. Saint Andrews, UK. 10 pp.
URL: http://www.maths.bath.ac.uk/~sw283/simon/dde.html
}

\seealso{ \code{\link{dde}} }

\keyword{ package }

